hdis_patient_management/requirements.txt
requests
python-dateutil
Django==4.0.1
django-mysql==4.4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
django-oauth-toolkit==2.3.0
mysqlclient
pika==1.2.0